API: Move Probe class to separate file and rename dx dt
The scan submodule was getting very long and scrolling was
annoying. Moved Probe to it's own file to shorten scrolls.

Rename dx dt to xstep tstep for clarity
